refactor: converter base module
